+++ src/core/shared/src/main/java/shared/service/results/GetBalancesResult.java
@@ -1,6 +1,7 @@
 package shared.service.results;
 
+import java.math.BigDecimal;
 import java.util.HashMap;
 
-public record GetBalancesResult(HashMap<String, Long> balances) {
+public record GetBalancesResult(HashMap<String, BigDecimal> balances) {
 }
