+++ src/core/webService/src/main/java/webService/api/models/OrderBookEntryModel.java
@@ -1,6 +1,8 @@
 package webService.api.models;
 
-public record OrderBookEntryModel(long size, long price, boolean canCancel, long id) {
+import java.math.BigDecimal;
+
+public record OrderBookEntryModel(BigDecimal size, BigDecimal price, boolean canCancel, long id) {
     public OrderBookEntryModel(long currentAccountId, shared.orderBook.OrderBookEntry entry) {
         this(entry.size(), entry.price(), entry.accountId() == currentAccountId, entry.id());
     }
